refactored structure of project and updated example project to support node v10
