+++ src/configs/env.ts
@@ -1,15 +1,11 @@
-const SPECIAL_TAGS = ['@debug', '@skip'];
+const RESERVED_TAGS = ['@debug', '@skip'];
 
 const All_TAGS = (process.env.TAGS || '')
     .split(',')
     .filter((tag) => tag)
     .map((tag) => tag.trim());
 
-export const TAGS = All_TAGS.filter((tag) => !SPECIAL_TAGS.includes(tag));
-
-export const HAS_SPECIAL_TAGS = All_TAGS.some((tag) =>
-    SPECIAL_TAGS.includes(tag)
-);
+export const TAGS = All_TAGS.filter((tag) => !RESERVED_TAGS.includes(tag));
 
 export const EXCLUDE_TAGS = TAGS.filter((tag) => tag.startsWith('not')).map(
     (tag) => tag.split('not ')[1]
