+++ src/configs/space.ts
@@ -1,0 +1,1 @@
+export const space = ' '.repeat(6);
