+++ src/utils/isJson.ts
@@ -1,0 +1,8 @@
+export function isJson(text): boolean {
+    try {
+        JSON.parse(text);
+        return true;
+    } catch (e) {
+        return false;
+    }
+}
